// frontend/apps/platform/src/enums/signupEnum.ts
export enum SignUpStep {
  SIGNUP = 'signup',
  VERIFY = 'verify',
  PASSWORD = 'password',
  ACCOUNT_TYPE = 'account-type',
  INDIVIDUAL_WORKSPACE = 'individual-workspace',
  WORKSPACE_TYPE = 'workspace-type',
  ACCESS_FORM = 'access-form',
  REQUEST_ACCESS = 'request-access',
  TELEGRAM_GROUP = 'telegram-group',
  LOADING = 'loading',
  REQUESTED_ACCESS = 'requested-access',
  ACCESS_REQUESTED = 'access-requested',
  CREATE_PROJECT = 'create-project',
}